Added reduced angle wall lights,  180EML and 90EML.  Use 180 for lengths of wall and 90 for corners.
The positions/rotations should be aligned such that setting the prefab's x & z position and y rotation equal to that of the wall will align the light with that wall.

For corners, position is at the corner itself. Rotation should be the same as the right side wall when facing the corner.